drf_knox/requirements/base.txt
# Django 
Django>=4.1,<4.3
django-kn-defaults==0.3.2
djvue==0.2.1
sentry-sdk==1.11.1
elastic-apm==6.13.2






# Django filer plugin
django-filer==2.2.3
easy-thumbnails==2.8.3


# for windows please add following lib before installing compressor
# rcssmin==1.1.0 --install-option="--without-c-extensions"
django-compressor==4.1


djangorestframework==3.14.0
django-environ==0.9.0




